vm: Extract label backpatching into SymbolTable.
Somewhat clearer, though I have a nagging doubt that each of these
classes and structs just obscure the intent of the code. It's unsettling
to code near the border between "small example program" and  "program
that may be extended later"; do the borders provide help when extending
or do they just obscure the intent of the code?