cryptopals: Use cmake as build system.
I've had a couple of hard-to-track-down bugs caused by grasping over to
much at once; writing several functions that depends on each other
without verifying the parts first.

I'll use gtest for unit testing; hoping that will make the verification
of each function more streamlined.

In preparation of that change, I'm replacing the Makefile with a
CMakeLists.txt file. I tend to do that - start off with a Makefile and
quickly realizing the limitations, replacing it with cmake.