oberon-0: Fix code generation for arrays
It's annoying to have to manually keep track of the stack of registers.
This took more time than I expected due to me mixing thinking that I had
the var_index in a certain stack slot when it was infact the index of
the array. Here's how the an address in the Oberon computer is
computed:

    mem = base_mem + var_index + (index * size)